Settings/Compare: Refactor module namespace once more, `settings` to top-level
