+++ cratedb_toolkit/settings/compare.py
